1.6.3 affichage réunions dans planning + fix annulation booking + equipement public

+++ src/java/net/algem/planning/BookingScheduleElement.java
@@ -30,9 +30,9 @@ public class BookingScheduleElement
   extends ScheduleElement
 {
 
-  private byte status;
+  private int status;
 
-  public byte getStatus() {
+  public int getStatus() {
     return status;
   }
